refactor: Split out the get/set/delete tests.
